+++ graph/assemblygraph.h
@@ -148,7 +148,7 @@ public:
 
     static QByteArray addNewlinesToSequence(QByteArray sequence, int interval = 70);
     int getDeadEndCount() const;
-    void getNodeStats(int * n50, int * shortestNode, int * firstQuartile, int * median, int * thirdQuartile, int * longestNode) const;
+    void getNodeStats(int * n50, int * shortestNode, double *firstQuartile, double *median, double *thirdQuartile, int * longestNode) const;
 
 private:
     template<typename T> double getValueUsingFractionalIndex(std::vector<T> * v, double index) const;
